# By g_nelson
# Via g_nelson
* commit 'be01cd9b3fdc2d73d7b9314f3765742adf677e9e':
  Implicitly cast the size to an int for 32-bit builds

+++ src/core/lib/core_data_model/simple_active_filters_model.cpp
@@ -102,7 +102,7 @@ const int & SimpleActiveFiltersModel::removeFilter() const
 void SimpleActiveFiltersModel::removeFilter( const int & index )
 {
 	impl_->removedIndex_ = index;
-	if (impl_->removedIndex_ == -1 || index >= impl_->filters_.size())
+	if (impl_->removedIndex_ == -1 || index >= (int) impl_->filters_.size())
 	{
 		return;
 	}
